move back delete assignments
